Fixed some wierd stuff, this looks clean and workd

+++ srcs/parser/parserUtils.cpp
@@ -23,9 +23,9 @@ std::vector<std::string> splitString(const std::string& input, char delimiter) {
 }
 
 
-unsigned long stringToUnsignedLong(const std::string& str) {
+short stringToUnsignedLong(const std::string& str) {
     std::istringstream stream(str);
-    unsigned long result = 0;
+    short result = 0;
 
     if (!(stream >> result) || !stream.eof()) {
         throw std::invalid_argument("Error en la conversión de cadena a unsigned long");
